+++ iam/auth/auth-user/application/authorize-oauth/authorize-oauth.result.ts
@@ -1,0 +1,6 @@
+import { IsUrl } from 'class-validator';
+
+export class AuthorizeOAuthResult {
+  @IsUrl()
+  authUrl: string;
+}
