+++ iam/auth/auth-user/application/logout/logout.command.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export class LogoutCommand {
+  @IsString()
+  @IsNotEmpty()
+  userId: string;
+}
