+++ iam/auth/auth-user/application/oauth-login/oauth-login.command.ts
@@ -1,0 +1,9 @@
+import { OAuthProviderType } from '../../domain/value-object/oauth-provider.enum';
+
+export class OAuthLoginCommand {
+  constructor(
+    public readonly oAuthProviderType: OAuthProviderType,
+    public readonly code: string,
+    public readonly state?: string,
+  ) {}
+}
