+++ iam/auth/auth-user/application/oauth-login/oauth-login.result.ts
@@ -1,0 +1,18 @@
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export class OAuthLoginResult {
+  @IsString()
+  @IsNotEmpty()
+  accessToken: string;
+
+  @IsString()
+  @IsNotEmpty()
+  refreshToken: string;
+
+  @IsString()
+  @IsNotEmpty()
+  userId: string;
+
+  @IsString()
+  redirectUrl: string;
+}
