+++ iam/auth/auth-user/application/renew-token/renew-token.command.ts
@@ -1,0 +1,11 @@
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export class RenewTokenCommand {
+  @IsString()
+  @IsNotEmpty()
+  userId: string;
+
+  @IsString()
+  @IsNotEmpty()
+  jti: string;
+}
