+++ iam/auth/auth-user/application/unlink-oauth/unlink-oauth.command.ts
@@ -1,0 +1,8 @@
+import { OAuthProviderType } from '../../domain/value-object/oauth-provider.enum';
+
+export class UnlinkOAuthCommand {
+  constructor(
+    public readonly oAuthProviderType: OAuthProviderType,
+    public readonly userId: string,
+  ) {}
+}
