+++ iam/auth/auth-user/domain/auth-user.repository.ts
@@ -1,0 +1,12 @@
+import { AuthUser } from './auth-user';
+
+export interface AuthUserStore {
+  save(authUser: AuthUser): Promise<void>;
+  update(authUser: AuthUser): Promise<void>;
+  findByOAuthIdandProvider(oauthId: string, provider: string): Promise<AuthUser | null>;
+  findByRefreshToken(refreshToken: string): Promise<AuthUser | null>;
+  findByUserId(userId: string): Promise<AuthUser | null>;
+  deleteById(id: string): Promise<void>;
+}
+
+export const AUTH_USER_STORE = Symbol('AUTH_USER_STORE');
