+++ iam/auth/auth-user/domain/event/auth-created.event.ts
@@ -1,0 +1,15 @@
+import { BaseDomainEvent } from 'src/shared/core/domain/base.domain-event';
+import { Identifier } from 'src/shared/core/domain/identifier';
+import { OAuthProviderType } from '../value-object/oauth-provider.enum';
+
+export class AuthCreatedEvent implements BaseDomainEvent {
+  readonly timestamp: Date;
+
+  constructor(
+    public readonly userId: Identifier,
+    public readonly email: string,
+    public readonly provider: OAuthProviderType,
+  ) {
+    this.timestamp = new Date();
+  }
+}
