+++ iam/auth/auth-user/domain/event/auth-deleted.event.ts
@@ -1,0 +1,9 @@
+import { BaseDomainEvent } from 'src/shared/core/domain/base.domain-event';
+
+export class AuthDeletedEvent implements BaseDomainEvent {
+  readonly timestamp: Date;
+
+  constructor(public readonly userId: string) {
+    this.timestamp = new Date();
+  }
+}
