--- src/auth/auth-user/application/authorize-oauth/authorize-oauth.use-case.ts
@@ -1,20 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { OAuthProviderFactory } from 'src/auth/core/infrastructure/oauth/oauth-provider.factory';
-import { AuthorizeOAuthRequestDto } from './dto/authorize-oauth.request.dto';
-import { AuthorizeOAuthResponseDto } from './dto/authorize-oauth.response.dto';
-
-@Injectable()
-export class AuthorizeOAuthUseCase {
-  constructor(private readonly oAuthProviderFactory: OAuthProviderFactory) {}
-
-  execute(requestDto: AuthorizeOAuthRequestDto): AuthorizeOAuthResponseDto {
-    const { oAuthProviderType, redirectUrl } = requestDto;
-    const provider = this.oAuthProviderFactory.getProvider(oAuthProviderType);
-
-    const encodedState = encodeURIComponent(redirectUrl || '');
-
-    const authUrl = provider.getAuthorizationUrl(encodedState);
-
-    return { authUrl };
-  }
-}
