--- src/auth/auth-user/application/authorize-oauth/dto/authorize-oauth.response.dto.ts
@@ -1,6 +1,0 @@
-import { IsUrl } from 'class-validator';
-
-export class AuthorizeOAuthResponseDto {
-  @IsUrl()
-  authUrl: string;
-}
