--- src/auth/auth-user/application/logout/dto/logout.request.dto.ts
@@ -1,7 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class LogoutRequestDto {
-  @IsString()
-  @IsNotEmpty()
-  userId: string;
-}
