--- src/auth/auth-user/application/logout/logout.use-case.ts
@@ -1,22 +1,0 @@
-import { Inject, Injectable } from '@nestjs/common';
-import { LogoutRequestDto } from './dto/logout.request.dto';
-import { CustomException } from 'src/shared/exception/custom-exception';
-import { CustomExceptionCode } from 'src/shared/exception/custom-exception-code';
-import { AUTH_USER_REPOSITORY, AuthUserRepository } from '../../domain/auth-user.repository';
-
-@Injectable()
-export class LogoutUseCase {
-  constructor(
-    @Inject(AUTH_USER_REPOSITORY)
-    private readonly authRepository: AuthUserRepository,
-  ) {}
-
-  async execute(requestDto: LogoutRequestDto): Promise<void> {
-    const { userId } = requestDto;
-    const auth = await this.authRepository.findByUserId(userId);
-    if (!auth) throw new CustomException(CustomExceptionCode.AUTH_INFO_NOT_FOUND);
-
-    auth.updateRefreshToken(null, new Date());
-    await this.authRepository.update(auth);
-  }
-}
