--- src/auth/auth-user/application/oauth-login/dto/oauth-login.response.dto.ts
@@ -1,18 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class OAuthLoginResponseDto {
-  @IsString()
-  @IsNotEmpty()
-  accessToken: string;
-
-  @IsString()
-  @IsNotEmpty()
-  refreshToken: string;
-
-  @IsString()
-  @IsNotEmpty()
-  userId: string;
-
-  @IsString()
-  redirectUrl: string;
-}
