--- src/auth/auth-user/application/oauth-login/oauth-login.command.ts
@@ -1,10 +1,0 @@
-import { ICommand } from '@nestjs/cqrs';
-import { OAuthProviderType } from 'src/auth/auth-user/domain/value-object/oauth-provider.enum';
-
-export class OAuthLoginCommand implements ICommand {
-  constructor(
-    public readonly oAuthProviderType: OAuthProviderType,
-    public readonly code: string,
-    public readonly state?: string,
-  ) {}
-}
