--- src/auth/auth-user/application/renew-token/dto/renew-token.request.dto.ts
@@ -1,11 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class RenewTokenRequestDto {
-  @IsString()
-  @IsNotEmpty()
-  userId: string;
-
-  @IsString()
-  @IsNotEmpty()
-  jti: string;
-}
