--- src/auth/auth-user/application/renew-token/dto/renew-token.response.dto.ts
@@ -1,11 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class RenewTokenResponseDto {
-  @IsString()
-  @IsNotEmpty()
-  accessToken: string;
-
-  @IsString()
-  @IsNotEmpty()
-  refreshToken: string;
-}
