--- src/auth/auth-user/application/unlink-oauth/unlink-oauth.command.ts
@@ -1,9 +1,0 @@
-import { ICommand } from '@nestjs/cqrs';
-import { OAuthProviderType } from '../../domain/value-object/oauth-provider.enum';
-
-export class UnlinkOAuthCommand implements ICommand {
-  constructor(
-    public readonly oAuthProviderType: OAuthProviderType,
-    public readonly userId: string,
-  ) {}
-}
