--- src/auth/auth-user/domain/event/login-succeeded.event.ts
@@ -1,14 +1,0 @@
-import { BaseDomainEvent } from 'src/shared/core/domain/base.domain-event';
-import { Identifier } from 'src/shared/core/domain/identifier';
-import { OAuthProviderType } from '../value-object/oauth-provider.enum';
-
-export class LoginSucceededEvent implements BaseDomainEvent {
-  readonly timesstamp: Date;
-
-  constructor(
-    public readonly userId: Identifier,
-    public readonly provider: OAuthProviderType,
-  ) {
-    this.timesstamp = new Date();
-  }
-}
